==> analysis/Build-Data/00_build-census.ipynb <==
# Prep Census Data

import geopandas as gpd
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm, trange
import json
storage = "/Volumes/easystore/Drones"

with open(f'{storage}/census/acs-block-groups-household-income/metadata.json') as f:
    acs_columns = json.load(f)
    
    
    

acs_poverty_bg = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"{storage}/census/acs-block-groups-household-income/acs2021_5yr_B19001_15000US060730134093.csv", chunksize=100000, dtype=str
            ),
            desc="Loading data",
        )
    ]
)

census_columns= pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"{storage}/census/race/DECENNIALDHC2020.P9-Column-Metadata.csv", chunksize=100000, dtype=str
            ),
            desc="Loading data",
        )
    ]
)
census_race= pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"{storage}/census/race/DECENNIALDHC2020.P9-Data.csv", chunksize=100000, dtype=str
            ),
            desc="Loading data",
        )
    ]
)
western_tracts= pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"../../data/manually-label/WesternTracts.csv", chunksize=100000, dtype=str
            ),
            desc="Loading data",
        )
    ]
)


## ACS Data
Free and Reduced Meals in Schools in California is 55k: https://www.cde.ca.gov/ls/nu/rs/scales2324.asp

columns = acs_columns['tables']['B19001']['columns']
remap={}
eligible = []
below_median = []
drop=[]
for column in columns:
    remap[column] = columns[column]['name'].replace(':','').replace('$','')
    remap[column+", Error"] = columns[column]['name'].replace(':','').replace('$','') + " Error"
    drop.append(columns[column]['name'].replace(':','').replace('$','') + " Error")
    
    if columns[column]['name'] != "Total:":
        if columns[column]['FRLP'] == "True":
            eligible.append(columns[column]['name'].replace(':','').replace('$',''))
        if columns[column]['median'] == "below":
            below_median.append(columns[column]['name'].replace(':','').replace('$',''))

# df.rename(columns={"A": "a", "B": "c"})
acs_poverty_bg = acs_poverty_bg.rename(columns=remap)

acs_poverty_bg['len'] = acs_poverty_bg['geoid'].apply(lambda x: len(x))
acs_poverty_bg = acs_poverty_bg[acs_poverty_bg['len'] == 19].copy()
acs_poverty_bg['block-group-geoid'] = acs_poverty_bg['geoid'].apply(lambda x: x.split('15000US')[1])


acs_poverty_bg['FRLP Households'] = acs_poverty_bg[eligible].astype(float).sum(axis=1)
acs_poverty_bg['Below Median Households'] = acs_poverty_bg[below_median].astype(float).sum(axis=1)
acs_poverty_bg['% FRLP Eligible'] = acs_poverty_bg['FRLP Households']/acs_poverty_bg['Total'].astype(float)
acs_poverty_bg['% Below Median'] = acs_poverty_bg['Below Median Households']/acs_poverty_bg['Total'].astype(float)
acs_poverty_bg = acs_poverty_bg.drop(labels=drop, axis=1)
acs_poverty_bg.head()

acs_poverty_bg.to_csv('../../data/outputs/outputs_acs-poverty-blockgroups.csv',index=False)

acs_poverty_bg[acs_poverty_bg['geoid']=='15000US060730032041'][['Less than 10,000', '10,000 to 14,999',
       '15,000 to 19,999', '20,000 to 24,999', '25,000 to 29,999',
       '30,000 to 34,999', '35,000 to 39,999', '40,000 to 44,999',
       '45,000 to 49,999', '50,000 to 59,999']].sum().astype(int).sum()



## Get Chula Vista Block Groups

cv_block_groups = acs_poverty_bg[['geoid','block-group-geoid']]
cv_block_groups.to_csv('../../data/outputs/outputs_cv-block-groups.csv',index=False)


## Census Data

columns = census_race.iloc[0].to_dict()
remap={}
drop=[]
for column in columns:
    try:
        if column[-1] == "A":
            drop.append(column)
        else:
            
            remap[column] = ' '.join(columns[column].split('!!')).replace(':',"").replace(' Hispanic or Latino',' Latino').replace('Not Latino Population of one race','').replace('alone','').replace('  ',' ').strip()
    except:
        print(f"Not Remapping: {column}")
        

census_race_remapped = census_race.rename(columns = remap)
census_race_remapped = census_race_remapped.drop(labels=drop, axis=1)
census_race_remapped=census_race_remapped[['Geography', 'Geographic Area Name', 'Total', 'Total Latino',
       'Total Not Latino', 'Total', 'Total White',
       'Total Black or African American',
       'Total American Indian and Alaska Native', 'Total Asian',
       'Total Native Hawaiian and Other Pacific Islander',
       'Total Some Other Race']][3:]

census_race_remapped.columns = ['Geography', 'Geographic Area Name', 'Total', 'Total Alone', 'Total Latino',
       'Total Not Latino', 'Total_useless', 'Total_useless_1', 'Total White',
       'Total Black or African American',
       'Total American Indian and Alaska Native', 'Total Asian',
       'Total Native Hawaiian and Other Pacific Islander',
       'Total Some Other Race']

census_race_remapped=census_race_remapped[['Geography', 'Geographic Area Name', 'Total', 'Total Alone', 'Total Latino',
       'Total Not Latino', 'Total White',
       'Total Black or African American',
       'Total American Indian and Alaska Native', 'Total Asian',
       'Total Native Hawaiian and Other Pacific Islander',
       'Total Some Other Race']]

census_race_remapped['GEOID20'] = census_race_remapped['Geography'].apply(lambda x: x.split('US')[-1])
census_race_remapped['len'] = census_race_remapped['GEOID20'].apply(lambda x: len(x))
mask=census_race_remapped['GEOID20'].str.contains('|'.join(cv_block_groups["block-group-geoid"].values), case=False)
census_race_remapped = census_race_remapped[mask].copy()

# census_race[census_race['GEOID20'].str.contains("060730032041")]

census_race_remapped.to_csv('../../data/outputs/outputs_census-race.csv',index=False)


sd_county = pd.read_csv('../../data/mapping/sd_county-geometry.csv')

sd_county['GEOID20'] = "0" + sd_county['GEOID20'].astype(str)
mask=sd_county['GEOID20'].astype(str).str.contains('|'.join(cv_block_groups["block-group-geoid"].values), case=False)
cv_blocks = sd_county[mask].copy()
cv_blocks['GEOID20'] = "0" + cv_blocks['GEOID20'].astype(int).astype(str)
cv_blocks['len'] = cv_blocks['GEOID20'].apply(lambda x: len(x))
cv_blocks['POP20'] = cv_blocks['POP20'].astype(int) + 1


cv_blocks['west_cv']=False

cv_blocks.loc[cv_blocks['TRACTCE20'].astype(str).isin(western_tracts['Tract'].astype(str)),'west_cv']=True

cv_blocks[cv_blocks['west_cv']==True]

cv_blocks.to_csv('../../data/outputs/outputs_cv-blocks-geometry.csv',index=False)

cv_blocks
