==> analysis/Build-Data/01_build-analysis-data.ipynb <==
# Build Analysis Data

import geopandas as gpd
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm, trange
from datetime import date
from multiprocess.pool import Pool


storage = "/Volumes/easystore/Drones/"
calls_for_service = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"{storage}/calls-for-service-with-census-data.csv",
                chunksize=100000,
                dtype=str,
            ),
            desc="Loading data",
        )
    ]
)
flight_details = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"{storage}/gpx-with-census-data.csv", 
                chunksize=100000, 
                dtype=str
            ),
            desc="Loading data",
        )
    ]
)
flights = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"../../data/all-flights-manifest.csv", 
                chunksize=100000, 
                dtype=str
            ),
            desc="Loading data",
        )
    ]
)

cv_blocks = pd.read_csv('../../data/outputs/outputs_cv-blocks-geometry.csv',dtype='str')

# flight_data = pd.read_csv("/Volumes/easystore/Drones/compiled-flight-data-gpx.csv",dtype=str)
# flight_data.head()

calls_for_service.shape[0]

cv_blocks.shape[0]

### Flights and Flight Details

cvpd_flights = flights[flights['department'] == 'cvpd'].copy()
cvpd_flights["date"] = pd.to_datetime(cvpd_flights["date"])

mask = (cvpd_flights['date'] >= '2021-07-01 00:00') & (cvpd_flights['date'] <= '2023-09-02 00:00')

cvpd_flights = cvpd_flights[mask].sort_values('date').copy()
cvpd_flights['clean-type'] = cvpd_flights['type'].str.lower()
cvpd_flights['clean-type'] = cvpd_flights['clean-type'].apply(lambda x: str(x).replace('  ','').strip())
cvpd_flights["match"] = cvpd_flights["incident_id"].astype(str).apply(lambda x: x.strip().replace(' ','').replace('L','0').replace('l','0')[-5:])

cvpd_flights.to_csv('../../data/outputs/outputs_cv-flight-manifest.csv',index=False)

flight_details["GEOID20"] = flight_details["geoid"].astype(str)
flight_details["len"] = flight_details["GEOID20"].apply(lambda x: len(x))
flight_details.loc[flight_details["len"] == 14, "GEOID20"] = "0" + flight_details["GEOID20"]
flight_details["len"] = flight_details["GEOID20"].apply(lambda x: len(x))


### Calls for Service

calls_for_service['len'] = calls_for_service['Full FIPS (block)'].astype(str).apply(lambda x: len(x))

calls_for_service['GEOID20'] = calls_for_service['Full FIPS (block)']
calls_for_service["match"] = (
    calls_for_service["Incident No."].astype(str).apply(lambda x: x.strip().replace(' ','').replace('l','0').replace('L','0')[-5:])
)
calls_for_service['clean-type'] = calls_for_service['Description'].str.lower()
calls_for_service['clean-type'] = calls_for_service['clean-type'].apply(lambda x: str(x).replace('  ','').strip())
calls_for_service.to_csv('../../data/outputs/outputs_cfs-data.csv',index=False)

calls_for_service.shape[0]

## Group by Block

### Call for Service

call_for_service_block_count = calls_for_service.groupby(["GEOID20"]).count()
call_for_service_block_count = call_for_service_block_count.reset_index()[
    ["GEOID20", "Incident No."]
]
call_for_service_block_count.columns = ["GEOID20", "call_count"]
call_for_service_block_count.head()

call_for_service_block_count_with_shps = pd.merge(
    cv_blocks, call_for_service_block_count, how="left", on=["GEOID20"]
)

call_for_service_block_count_with_shps["GEOID20"].apply(
    lambda x: len(x)
).drop_duplicates()


call_for_service_block_count_with_shps[
    "call_count"
] = call_for_service_block_count_with_shps["call_count"].replace(np.nan, 0)

call_for_service_block_count_with_shps['weight'] = call_for_service_block_count_with_shps['call_count'].astype(int)/call_for_service_block_count_with_shps['POP20'].astype(int)

call_for_service_block_count_with_shps[
    "weight"
] = call_for_service_block_count_with_shps["weight"].replace(np.nan, 0)
call_for_service_block_count_with_shps[
    "weight"
] = call_for_service_block_count_with_shps["weight"].replace(np.inf, 0)


call_for_service_block_count_with_shps.to_csv('../../data/outputs/outputs_cfs-data-with-shps.csv',index=False)

### Drone Seconds in Block

flight_details = flight_details[flight_details['id'].isin(cvpd_flights['id'])].copy()

# Strip the millisecond component from the 'sequence' timestamps
flight_details['sequence'] = pd.to_datetime(flight_details['sequence'])
flight_details['sequence_seconds'] = flight_details['sequence'].dt.floor('S')

# Now count the number of unique seconds in each block
unique_seconds_per_block_seconds = flight_details.groupby('GEOID20')['sequence_seconds'].nunique()

unique_seconds_per_block_seconds.head()


unique_seconds_per_block_seconds = pd.DataFrame(unique_seconds_per_block_seconds).reset_index()

unique_seconds_per_block_seconds.columns = ['GEOID20','seconds']

unique_seconds_per_block_seconds

unique_seconds_in_block_with_shps = pd.merge(
    cv_blocks, unique_seconds_per_block_seconds, how="left", on=["GEOID20"]
)
unique_seconds_in_block_with_shps["seconds"] = unique_seconds_in_block_with_shps[
    "seconds"
].replace(np.nan, 0)


unique_seconds_in_block_with_shps['ALAND20'] = unique_seconds_in_block_with_shps['ALAND20'].astype(int)

unique_seconds_in_block_with_shps["weight"] = (
    unique_seconds_in_block_with_shps["seconds"]
    / unique_seconds_in_block_with_shps["ALAND20"]
)


unique_seconds_in_block_with_shps = pd.merge(unique_seconds_in_block_with_shps,call_for_service_block_count_with_shps[['GEOID20','call_count']],how='left',on='GEOID20')

unique_seconds_in_block_with_shps.to_csv('../../data/outputs/outputs_drone-seconds-block-count.csv',index=False)

### Drone Days / CFS in Block
- 792 days in analysis

from datetime import date
 
def numOfDays(date1, date2):
    if date2 > date1:   
        return (date2-date1).days
    else:
        return (date1-date2).days

start= date(2021, 7, 1)
end = date(2023, 9, 1)

days_in_analysis = numOfDays(start,end)

def calculate_summary(index):
    block = cv_blocks.iloc[index]
    drones_on_block = flight_details[flight_details['GEOID20']==block['GEOID20']] 
    calls_to_block = calls_for_service[calls_for_service['GEOID20']==block['GEOID20']] 

    block['num_flights'] = drones_on_block['id'].drop_duplicates().shape[0]
    block['flights_per_day'] = block['num_flights']/days_in_analysis
    block['num_calls'] = calls_to_block.shape[0]
    block['calls_per_day'] = block['num_calls']/days_in_analysis
    return block

from multiprocess.pool import Pool

with Pool(10) as pool:

    drones_cfs_pdpb = list(
        tqdm(pool.imap(calculate_summary, range(0, cv_blocks.shape[0])), total=cv_blocks.shape[0])
    )
    drones_cfs_pdpb = pd.DataFrame(drones_cfs_pdpb)

drones_cfs_pdpb.to_csv('../../data/outputs/output_cv-blocks-drones-calls.csv',index=False)

drones_cfs_pdpb.sort_values('num_flights').tail()

unique_days_in_block
