==> analysis/Build-Data/01_types-of-flights.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm, trange
import re

storage = "/Volumes/easystore/Drones/"

cv_flights = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"../../data/outputs/outputs_cv-flight-manifest.csv", chunksize=100000, dtype=str
            ),
            desc="Loading data",
        )
    ]
)
calls_for_service = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"{storage}/calls-for-service-with-census-data.csv",
                chunksize=100000,
                dtype=str,
            ),
            desc="Loading data",
        )
    ]
)

cv_flights['clean-type'] = cv_flights['type'].str.lower()
cv_flights['clean-type'] = cv_flights['clean-type'].apply(lambda x: str(x).replace('  ','').strip())

cv_flights[cv_flights['clean-type'].astype(str).str.contains('vaping')]

cv_flights['clean-type'].drop_duplicates().to_csv('../../data/manually-label/types-to-label-flights.csv',index=False)


cv_flights[cv_flights['clean-type'].str.contains('drone')]
cv_flights['src']='drone_flights'

calls_for_service['clean-type'] = calls_for_service['Description'].str.lower()
calls_for_service['src']='cfs'
calls_for_service['clean-type'] = calls_for_service['clean-type'].apply(lambda x: str(x).replace('  ','').strip())

a = []
a.append(calls_for_service[['src','clean-type']])
a.append(cv_flights[['src','clean-type']])
types = pd.concat(a)

types[['clean-type']].drop_duplicates()

types[['clean-type']].drop_duplicates().to_csv('../../data/manually-label/types.csv',index=False)
