==> analysis/Build-Data/03_sensitive-locations.ipynb <==
import os

import fiona
import geopandas as gpd
import gpxpy
import numpy as np
import pandas as pd
from multiprocess.pool import Pool
from tqdm.notebook import tqdm, trange
import warnings
import geopy.distance

warnings.filterwarnings('ignore')

fiona.drvsupport.supported_drivers['kml'] = 'rw'
fiona.drvsupport.supported_drivers['KML'] = 'rw'

storage = "/Volumes/easystore/Drones/"

cv_flights = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"../../data/outputs/outputs_cv-flight-manifest.csv", chunksize=100000, dtype=str
            ),
            desc="Loading data",
        )
    ]
)
per_second = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"../../data/outputs/outputs_drone-seconds-block-count.csv",
                chunksize=100000,
                dtype=str,
            ),
            desc="Loading data",
        )
    ]
)

flight_details = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"{storage}/gpx-with-census-data.csv", chunksize=100000, dtype=str
            ),
            desc="Loading data",
        )
    ]
)
locations = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"../../data/manually-label/sensitive-locations.csv", chunksize=100000, dtype=str
            ),
            desc="Loading data",
        )
    ]
)

locations['lat'] = locations['Coordinates'].apply(lambda x: x.split(',')[0].strip())
locations['lon'] = locations['Coordinates'].apply(lambda x: x.split(',')[1].strip())

flight_details = flight_details[flight_details['id'].isin(cv_flights['id'])].copy()

locations 

def calculate_distance(row, target_lat, target_lon):
    origin = (float(row['latitude']), float(row['longitude']))
    target = (float(target_lat), float(target_lon))
    distance_in_meters = geopy.distance.geodesic(origin, target).meters
    return distance_in_meters * 3.28084  # Convert meters to feet
# Apply the function to each row in the DataFrame



def find_sensitive_flights(index):

    meta = {}
    flight_id = flight_ids.iloc[index]['id']
    flight = flight_details[flight_details['id'] == flight_id]
    meta['flight_id'] = flight_id
    for index, t_row in locations.iterrows():
        flight[f"close_to_{t_row['Location']}"] = False
        flight[f"distance_to_{t_row['Location']}"] = flight.apply(calculate_distance, axis=1, args=(t_row['lat'],t_row['lon']))
        
        flight.loc[flight[f"distance_to_{t_row['Location']}"]<=50,f"close_to_{t_row['Location']}"] = True
        
        if flight[flight[f"close_to_{t_row['Location']}"] == True].shape[0]>0:
            meta[f"close_to_{t_row['Location']}"] = True
        else:
            meta[f"close_to_{t_row['Location']}"] = False
    
    return  meta

flight_ids = flight_details[['id']].drop_duplicates()
with Pool(50) as pool:

    lingering_nd = list(
        tqdm(pool.imap(find_sensitive_flights, range(0, flight_ids.shape[0])), total=flight_ids.shape[0])
    )
    lingering_nd = pd.DataFrame(lingering_nd)

lingering_nd.to_csv('./sensitive-flights-do-not-delete.csv')

lingering_nd.columns.values

lingering_nd[lingering_nd['close_to_USCIS San Diego ']==True]

geopy.distance.geodesic((flight_details['latitude'].values[0],flight_details['longitude'].values[0]),(32.60769923913585,-117.06656860342576)).miles

possible_cfs_by_distance = []
for _, flight in tqdm(orphan_flights.iterrows()):
#   find possible calls by time
    possible_cfs_for_flight = possible_calls_by_time_alone[possible_calls_by_time_alone['drone-id']==flight['id']]
    
    orphan_flight = orphan_flight_blocks[orphan_flight_blocks['id']==flight['id']].copy()
    orphan_flight['seconds'] = orphan_flight['seconds'].astype(int)

    lingering_blocks = orphan_flight[orphan_flight['seconds']>=60]
    lingering_blocks = cv_blocks[cv_blocks['GEOID20'].isin(lingering_blocks['GEOID20'])]
    if lingering_blocks.shape[0]>0:
        for _, cfs in possible_cfs_for_flight.iterrows():
            for _, block in lingering_blocks.iterrows():
                distance = geopy.distance.geodesic((cfs['Latitude'],cfs['Longitude']),(block['INTPTLAT20'],block['INTPTLON20'])).miles
                if str(cfs['Latitude'])=='0':
                    cfs["lingering-block"] = block['GEOID20']
                    cfs["distance"] = distance
                    possible_cfs_by_distance.append(cfs)

                elif distance <= .5:
                    cfs["lingering-block"] = block['GEOID20']
                    cfs["distance"] = distance
                    possible_cfs_by_distance.append(cfs)
                elif distance > 30:
                    cfs["lingering-block"] = block['GEOID20']
                    cfs["distance"] = distance
                    possible_cfs_by_distance.append(cfs)


def build_paths(array_of_ids):
    compiled_flights =[] 
    for flight_id in array_of_ids:

        flight_path = gpd.read_file(f"/Volumes/easystore/Drones/flights/kml/{flight_id}.kml",driver='KML')
        flight_path['id'] = flight_id
        compiled_flights.append(flight_path)
    return pd.concat(compiled_flights)

geocodio_flights = pd.concat(
    [
        chunk
        for chunk in tqdm(
            pd.read_csv(
                f"{storage}/geocodio/all-flights-manifest_geocodio.csv",
                chunksize=100000,
                dtype=str,
            ),
            desc="Loading data",
        )
    ]
)
# a = pd.read_csv('../../data/outputs/outputs_cv-flight-manifest.csv')

geocodio_flights[(geocodio_flights['address_map'].astype(str).str.contains('1700',case=False))&(geocodio_flights['address_map'].astype(str).str.contains('broadway',case=False))]

# pool = 060730123021012


060730127003001

lede = per_flight_block_seconds[(per_flight_block_seconds['GEOID20']=='060730132063003')]
flights = []
for f in lede['id'].drop_duplicates().values:
    flight = per_flight_block_seconds[per_flight_block_seconds['id']==f]
    flight = flight[flight['GEOID20']=='060730127003001']
    if flight.shape[0]>0:
        flights.append(flight)

pp_paths = build_paths(['29b1f6e179d9564bfac69b9a187c7ab6'])
pp_paths = pd.merge(pp_paths,geocodio_flights,how='left',on='id')

pp_paths[['id','geometry','lat_map','lon_map','type']].to_csv('./public_pool.csv',index=False)
# pp_paths[['geometry','id']].drop_duplicates().to_csv('./low-paths.csv',index=False)

flight_details[flight_details['id']=='4342f9e237e48c34ddd15271a0ed1dc9'][['sequence','latitude','longitude','altitude']].to_csv('./flight.csv',index=False)

pp_paths
